server/requirements.txt
# server/requirements.txt

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.1

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis for caching and sessions
redis==5.0.1
aioredis==2.0.1

# WebSocket support
websockets==12.0
aiofiles==23.2.1
aiohttp==3.9.1

# Security
cryptography==41.0.7
pycryptodome==3.19.0
pyotp==2.9.0

# DNS server functionality
dnslib==0.9.23
dnspython==2.4.2

# Task queue
celery==5.3.4
kombu==5.3.4

# Monitoring and logging
prometheus-client==0.19.0
python-json-logger==2.0.7
structlog==23.2.0

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# CLI tools
click==8.1.7
rich==13.7.0
typer==0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# Development
python-dotenv==1.0.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
pyyaml==6.0.1
pendulum==2.1.2
psutil==5.9.6
netifaces==0.11.0
